Shorten introduction in virtual-environments.rst
Graham, thank you so much for your work on mod_wsgi. I discovered your Paypal link on the Contributing page and have used it. Your efforts are hugely appreciated.

I love developer tools with great documentation. The docs for mod_wsgi are thorough (great!) but not always well-organized. Often, there is so much in places that it becomes hard to find whats important. I'm proposing a change here as an example.

The first sentence can be cut because the work it does is handled by the title of the page.

The second paragraph can be cut because the audience of this page likely knows the purpose of virtual environments. If they don’t they can look it up elsewhere.

The book On Writing Well by William Zinsser, specifically “Part I Principles” that includes chapters titled “Simplicity”, “Clutter”, “Style”, and “Words” was tremendously helpful to me and a text I’d recommend. A quote from it: “The secret of good writing is to strip every sentence to its cleanest components.”