Add options to the Django runmodwsgi command for use when binding to non-standard ports that the standard ports have been redirected to.
Prior to this feature if you set the http or https ports to any value other than the standard ports (80 and 443 respectively) then the port number is automatically added to the URL. While this behavior can be correct, it isn't always necessary or desirable. For instance, if the standard ports are forwarded and Apache is run bound to those forwarded ports, then the port doesn't need to be included in the URL and doing so may confuse end users. It can be desirable for many reasons to forward the standard ports, such as not running Apache as the root user just so that it can bind to the standard ports.

To address this, two options were added to the runmodwsgi command:

--port-redirect-standard
--https-redirect-standard

So for instance if you have port 80 redirected to port 8080 and port 443 redirected to port 8443 you would use the options like this:

${python} ${path_to_django_project}/manage.py runmodwsgi <....other options you use...> --port=8080 --port-redirect-standard --https-port=8443 --https-redirect-standard

Now your server will listen on the port and https-port you specified but will not add those ports to the URL.

Both options can be omitted or included independantly.